REF: Split OpenMM logic into smaller files
