+++ openff/interchange/interop/openmm/_nonbonded.py
@@ -1,0 +1,492 @@
+from typing import TYPE_CHECKING, Dict, Tuple
+
+import openmm
+from openff.units import unit as off_unit
+from openff.units.openmm import to_openmm as to_openmm_quantity
+from openmm import unit
+
+from openff.interchange.constants import _PME
+from openff.interchange.exceptions import (
+    InternalInconsistencyError,
+    UnsupportedCutoffMethodError,
+    UnsupportedExportError,
+)
+from openff.interchange.interop.parmed import _lj_params_from_potential
+from openff.interchange.models import TopologyKey
+
+if TYPE_CHECKING:
+    from openff.interchange import Interchange
+
+_LORENTZ_BERTHELOT = "sigma=(sigma1+sigma2)/2; epsilon=sqrt(epsilon1*epsilon2); "
+
+
+def _process_nonbonded_forces(openff_sys, openmm_sys, combine_nonbonded_forces=False):
+    """
+    Process the non-bonded handlers in an Interchange into corresponding openmm objects.
+
+    This typically involves processing the vdW and Electrostatics sections of an Interchange object
+    into a corresponding openmm.NonbondedForce (if `combine_nonbonded_forces=True`) or a
+    collection of other forces (NonbondedForce, CustomNonbondedForce, CustomBondForce) if
+    `combine_nonbondoed_forces=False`.
+
+    """
+    from openff.interchange.components.smirnoff import _SMIRNOFFNonbondedHandler
+
+    for handler in openff_sys.handlers.values():
+        if isinstance(handler, _SMIRNOFFNonbondedHandler):
+            break
+    else:
+        return
+
+    has_virtual_sites = "VirtualSites" in openff_sys.handlers
+
+    if has_virtual_sites:
+        from openff.interchange.interop.openmm._virtual_sites import (
+            _check_virtual_site_exclusion_policy,
+        )
+
+        virtual_site_handler = openff_sys["VirtualSites"]
+
+        _check_virtual_site_exclusion_policy(virtual_site_handler)
+
+    # TODO: Process ElectrostaticsHandler.exception_potential
+    if "vdW" in openff_sys.handlers or "Electrostatics" in openff_sys.handlers:
+
+        _data = _prepare_input_data(openff_sys)
+
+        if combine_nonbonded_forces:
+            non_bonded_force = _create_single_nonbonded_force(_data, openff_sys)
+            openmm_sys.addForce(non_bonded_force)
+        else:
+            (
+                vdw_force,
+                electrostatics_force,
+                vdw_14_force,
+                coul_14_force,
+            ) = _create_multiple_nonbonded_forces(
+                _data,
+                openff_sys,
+            )
+            for force in [
+                vdw_force,
+                electrostatics_force,
+                vdw_14_force,
+                coul_14_force,
+            ]:
+                openmm_sys.addForce(force)
+
+    elif "Buckingham-6" in openff_sys.handlers:
+        buck_handler = openff_sys["Buckingham-6"]
+
+        non_bonded_force = openmm.CustomNonbondedForce(
+            "A * exp(-B * r) - C * r ^ -6; A = sqrt(A1 * A2); B = 2 / (1 / B1 + 1 / B2); C = sqrt(C1 * C2)"
+        )
+        non_bonded_force.addPerParticleParameter("A")
+        non_bonded_force.addPerParticleParameter("B")
+        non_bonded_force.addPerParticleParameter("C")
+        openmm_sys.addForce(non_bonded_force)
+
+        for _ in openff_sys.topology.atoms:
+            non_bonded_force.addParticle([0.0, 0.0, 0.0])
+
+        if openff_sys.box is None:
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.NoCutoff)
+        else:
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.CutoffPeriodic)
+            non_bonded_force.setCutoffDistance(buck_handler.cutoff * unit.angstrom)
+
+        for top_key, pot_key in buck_handler.slot_map.items():
+            atom_idx = top_key.atom_indices[0]
+
+            # TODO: Add electrostatics
+            params = buck_handler.potentials[pot_key].parameters
+            a = to_openmm_quantity(params["A"])
+            b = to_openmm_quantity(params["B"])
+            c = to_openmm_quantity(params["C"])
+            non_bonded_force.setParticleParameters(atom_idx, [a, b, c])
+
+        return
+
+    else:
+        # Here we assume there are no vdW interactions in any handlers
+        # vdw_handler = None
+
+        try:
+            electrostatics_handler = openff_sys["Electrostatics"]
+        except LookupError:
+            raise InternalInconsistencyError(
+                "In a confused state, could not find any vdW interactions but also failed to find "
+                "any electrostatics handler. This is a supported use case but should have been caught "
+                "earlier in this function. Please file an issue with a minimal reproducing example."
+            )
+
+        electrostatics_method = (
+            electrostatics_handler.periodic_potential
+            if electrostatics_handler
+            else None
+        )
+
+        non_bonded_force = openmm.NonbondedForce()
+        openmm_sys.addForce(non_bonded_force)
+
+        for _ in openff_sys.topology.atoms:
+            non_bonded_force.addParticle(0.0, 1.0, 0.0)
+
+        if electrostatics_method in ["Coulomb", None]:
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.NoCutoff)
+            # TODO: Would setting the dispersion correction here have any impact?
+        elif electrostatics_method == _PME:
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.LJPME)
+            non_bonded_force.setEwaldErrorTolerance(1.0e-4)
+        else:
+            raise UnsupportedCutoffMethodError(
+                f"Found no vdW interactions but an electrostatics method of {electrostatics_method}. "
+                "This is either unsupported or ambiguous. If you believe this exception has been raised "
+                "in error, please file an issue with a minimally reproducing example and your motivation "
+                "for this use case."
+            )
+
+
+def _prepare_input_data(openff_sys):
+    try:
+        vdw_handler = openff_sys["vdW"]
+    except LookupError:
+        vdw_handler = None
+
+    if vdw_handler:
+        vdw_cutoff = vdw_handler.cutoff
+        vdw_method = vdw_handler.method.lower()
+        mixing_rule = getattr(vdw_handler, "mixing_rule", None)
+        vdw_expression = vdw_handler.expression
+        vdw_expression = vdw_expression.replace("**", "^")
+    else:
+        vdw_cutoff = None
+        vdw_method = None
+        mixing_rule = None
+        vdw_expression = "(no handler found)"
+
+    try:
+        electrostatics_handler = openff_sys["Electrostatics"]
+    except LookupError:
+        electrostatics_handler = None
+
+    electrostatics_method = (
+        electrostatics_handler.periodic_potential if electrostatics_handler else None
+    )
+
+    return {
+        "vdw_handler": vdw_handler,
+        "vdw_cutoff": vdw_cutoff,
+        "vdw_method": vdw_method,
+        "vdw_expression": vdw_expression,
+        "mixing_rule": mixing_rule,
+        "mixing_rule_expression": _LORENTZ_BERTHELOT,
+        "electrostatics_handler": electrostatics_handler,
+        "electrostatics_method": electrostatics_method,
+        "periodic": openff_sys.box is None,
+    }
+
+
+def _create_single_nonbonded_force(
+    data: Dict,
+    openff_sys: "Interchange",
+) -> openmm.NonbondedForce:
+    """Create a single openmm.NonbondedForce from vdW/electrostatics/virtual site handlers."""
+    if data["mixing_rule"] not in ("lorentz-berthelot", None):
+        raise UnsupportedExportError(
+            "OpenMM's default NonbondedForce only supports Lorentz-Berthelot mixing rules."
+            "Try setting `combine_nonbonded_forces=False`."
+        )
+
+    non_bonded_force = openmm.NonbondedForce()
+
+    for _ in openff_sys.topology.atoms:
+        non_bonded_force.addParticle(0.0, 1.0, 0.0)
+
+    if openff_sys.box is None:
+        if (data["vdw_method"] in ("cutoff", None)) and (
+            data["electrostatics_method"] in ("Coulomb", None)
+        ):
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.NoCutoff)
+            non_bonded_force.setUseDispersionCorrection(True)
+            if data["vdw_cutoff"]:
+                non_bonded_force.setCutoffDistance(
+                    to_openmm_quantity(data["vdw_cutoff"])
+                )
+        else:
+            raise UnsupportedCutoffMethodError(
+                f"Combination of non-bonded cutoff methods {data['vdw_method']} (vdW) and "
+                f"{data['electrostatics_method']} (Electrostatics) not currently supported or "
+                f"invalid with `combine_nonbonded_forces=True` and `.box={openff_sys.box}`."
+            )
+
+    else:
+        if data["vdw_method"] in ("cutoff", None) and data["electrostatics_method"] in (
+            _PME,
+            None,
+        ):
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.PME)
+            non_bonded_force.setEwaldErrorTolerance(1.0e-4)
+            non_bonded_force.setUseDispersionCorrection(True)
+            if not data["vdw_cutoff"]:
+                # With no vdW handler and/or ambiguous cutoff, cannot set it,
+                # thereforce silently fall back to OpenMM's default. It's not
+                # clear if this value matters with only (PME) charges and no
+                # vdW interactions in the system.
+                pass
+            else:
+                non_bonded_force.setCutoffDistance(
+                    to_openmm_quantity(data["vdw_cutoff"])
+                )
+        elif data["vdw_method"] == "pme" and data["electrostatics_method"] == _PME:
+            non_bonded_force.setNonbondedMethod(openmm.NonbondedForce.LJPME)
+            non_bonded_force.setEwaldErrorTolerance(1.0e-4)
+        else:
+            raise UnsupportedCutoffMethodError(
+                f"Combination of non-bonded cutoff methods {data['vdw_method']} (vdW) and "
+                "{data['electrostatics_method']} (Electrostatics) not currently supported or invalid with "
+                f"`combine_nonbonded_forces=True` and `.box={openff_sys.box}`."
+            )
+
+    if data["electrostatics_handler"] is not None:
+        try:
+            partial_charges = data["electrostatics_handler"].charges_with_virtual_sites
+        except AttributeError:
+            partial_charges = data["electrostatics_handler"].charges
+
+    for atom_index, _ in enumerate(openff_sys.topology.atoms):
+        # TODO: Actually process virtual site vdW parameters here
+
+        top_key = TopologyKey(atom_indices=(atom_index,))
+
+        if data["electrostatics_handler"] is not None:
+            partial_charge = partial_charges[top_key].m_as(off_unit.e)
+        else:
+            partial_charge = 0.0
+
+        if data["vdw_handler"] is not None:
+            pot_key = data["vdw_handler"].slot_map[top_key]
+            sigma, epsilon = _lj_params_from_potential(
+                data["vdw_handler"].potentials[pot_key]
+            )
+            sigma = sigma.m_as(off_unit.nanometer)
+            epsilon = epsilon.m_as(off_unit.kilojoule / off_unit.mol)
+        else:
+            sigma = unit.Quantity(0.0, unit.nanometer)
+            epsilon = unit.Quantity(0.0, unit.kilojoules_per_mole)
+
+        non_bonded_force.setParticleParameters(
+            atom_index,
+            partial_charge,
+            sigma,
+            epsilon,
+        )
+
+    bonds = [
+        sorted(openff_sys.topology.atom_index(a) for a in bond.atoms)
+        for bond in openff_sys.topology.bonds
+    ]
+
+    coul_14 = (
+        data["electrostatics_handler"].scale_14
+        if "electrostatics_handler" in data
+        else 1.0
+    )
+    vdw_14 = data["vdw_handler"].scale_14 if "vdw_handler" in data else 1.0
+
+    non_bonded_force.createExceptionsFromBonds(
+        bonds=bonds,
+        coulomb14Scale=coul_14,
+        lj14Scale=vdw_14,
+    )
+
+    _apply_switching_function(data["vdw_handler"], non_bonded_force)
+
+    return non_bonded_force
+
+
+def _create_multiple_nonbonded_forces(
+    data: Dict,
+    openff_sys: "Interchange",
+) -> Tuple[
+    openmm.NonbondedForce,
+    openmm.CustomNonbondedForce,
+    openmm.CustomBondForce,
+    openmm.CustomBondForce,
+]:
+    vdw_force = openmm.CustomNonbondedForce(
+        data["vdw_expression"] + "; " + data["mixing_rule_expression"]
+    )
+    vdw_force.addPerParticleParameter("sigma")
+    vdw_force.addPerParticleParameter("epsilon")
+
+    for _ in openff_sys.topology.atoms:
+        vdw_force.addParticle([1.0, 0.0])
+
+    if data["vdw_method"] == "cutoff":
+        if openff_sys.box is None:
+            vdw_force.setNonbondedMethod(openmm.NonbondedForce.CutoffNonPeriodic)
+        else:
+            vdw_force.setNonbondedMethod(openmm.NonbondedForce.CutoffPeriodic)
+        vdw_force.setUseLongRangeCorrection(True)
+        vdw_force.setCutoffDistance(to_openmm_quantity(data["vdw_cutoff"]))
+
+        _apply_switching_function(data["vdw_handler"], vdw_force)
+
+    elif data["vdw_method"] == "pme":
+        if openff_sys.box is None:
+            raise UnsupportedCutoffMethodError(
+                "vdW method pme/ljpme is not valid for non-periodic systems."
+            )
+        else:
+            raise UnsupportedCutoffMethodError(
+                "LJ-PME with split non-bonded forces is not supported due to openmm.CustomNonbondedForce "
+                "not supporting PME. If also using PME electrostatics, try `combine_nonbonded_forces=True`,  "
+                "which should produce a single force with NonbondedForce.LJPME, which uses PME for both "
+                "electrostatics and LJ forces tersm. If your use case would benenfit from split non-bonded "
+                "forces with LJPME, please file an feature request."
+            )
+
+    electrostatics_force = openmm.NonbondedForce()
+
+    for _ in openff_sys.topology.atoms:
+        electrostatics_force.addParticle(0.0, 1.0, 0.0)
+
+    if data["electrostatics_method"] == "reaction-field":
+        raise UnsupportedExportError(
+            "Reaction field electrostatics not supported. If this use case is important to you, "
+            "please raise an issue describing the scope of functionality you desired."
+        )
+
+    elif data["electrostatics_method"] == _PME:
+        electrostatics_force.setNonbondedMethod(openmm.NonbondedForce.PME)
+        electrostatics_force.setEwaldErrorTolerance(1.0e-4)
+        electrostatics_force.setUseDispersionCorrection(True)
+        if data["vdw_cutoff"] is not None:
+            # All nonbonded forces must use the same cutoff, even though PME doesn't have a cutoff
+            electrostatics_force.setCutoffDistance(
+                to_openmm_quantity(data["vdw_cutoff"])
+            )
+    elif ["electrostatics_method"] == "cutoff":
+        raise UnsupportedCutoffMethodError(
+            "OpenMM does not clearly support cut-off electrostatics with no reaction-field attenuation."
+        )
+    else:
+        raise UnsupportedCutoffMethodError(
+            f"Electrostatics method {data['electrostatics_method']} not supported"
+        )
+
+    if data["electrostatics_handler"] is not None:
+        try:
+            partial_charges = data["electrostatics_handler"].charges_with_virtual_sites
+        except AttributeError:
+            partial_charges = data["electrostatics_handler"].charges
+
+    for atom_index, _ in enumerate(openff_sys.topology.atoms):
+        # TODO: Actually process virtual site vdW parameters here
+
+        top_key = TopologyKey(atom_indices=(atom_index,))
+
+        if data["electrostatics_handler"] is not None:
+            partial_charge = partial_charges[top_key].m_as(off_unit.e)
+        else:
+            partial_charge = 0.0
+
+        if data["vdw_handler"] is not None:
+            pot_key = data["vdw_handler"].slot_map[top_key]
+            sigma, epsilon = _lj_params_from_potential(
+                data["vdw_handler"].potentials[pot_key]
+            )
+            sigma = sigma.m_as(off_unit.nanometer)
+            epsilon = epsilon.m_as(off_unit.kilojoule / off_unit.mol)
+        else:
+            sigma = unit.Quantity(0.0, unit.nanometer)
+            epsilon = unit.Quantity(0.0, unit.kilojoules_per_mole)
+
+        vdw_force.setParticleParameters(atom_index, [sigma, epsilon])
+        electrostatics_force.setParticleParameters(atom_index, partial_charge, 0.0, 0.0)
+
+    # Attempting to match the value used internally by OpenMM; The source of this value is likely
+    # https://github.com/openmm/openmm/issues/1149#issuecomment-250299854
+    # 1 / * (4pi * eps0) * elementary_charge ** 2 / nanometer ** 2
+    coul_const = 138.935456  # kJ/nm
+
+    vdw_14_force = openmm.CustomBondForce("4*epsilon*((sigma/r)^12-(sigma/r)^6)")
+    vdw_14_force.addPerBondParameter("sigma")
+    vdw_14_force.addPerBondParameter("epsilon")
+    vdw_14_force.setUsesPeriodicBoundaryConditions(True)
+    coul_14_force = openmm.CustomBondForce(f"{coul_const}*qq/r")
+    coul_14_force.addPerBondParameter("qq")
+    coul_14_force.setUsesPeriodicBoundaryConditions(True)
+
+    bonds = [
+        sorted(openff_sys.topology.atom_index(a) for a in bond.atoms)
+        for bond in openff_sys.topology.bonds
+    ]
+
+    coul_14 = (
+        data["electrostatics_handler"].scale_14
+        if "electrostatics_handler" in data
+        else 1.0
+    )
+    vdw_14 = data["vdw_handler"].scale_14 if "vdw_handler" in data else 1.0
+
+    electrostatics_force.createExceptionsFromBonds(
+        bonds=bonds,
+        coulomb14Scale=coul_14,
+        lj14Scale=vdw_14,
+    )
+
+    for i in range(electrostatics_force.getNumExceptions()):
+        (p1, p2, q, sig, eps) = electrostatics_force.getExceptionParameters(i)
+
+        # If the interactions are both zero, assume this is a 1-2 or 1-3 interaction
+        if q._value == 0 and eps._value == 0:
+            pass
+        else:
+            # Assume this is a 1-4 interaction
+            # Look up the vdW parameters for each particle
+            sig1, eps1 = vdw_force.getParticleParameters(p1)
+            sig2, eps2 = vdw_force.getParticleParameters(p2)
+            q1, _, _ = electrostatics_force.getParticleParameters(p1)
+            q2, _, _ = electrostatics_force.getParticleParameters(p2)
+
+            # manually compute and set the 1-4 interactions
+            sig_14 = (sig1 + sig2) * 0.5
+            eps_14 = (eps1 * eps2) ** 0.5 * vdw_14
+            qq = q1 * q2 * coul_14
+
+            vdw_14_force.addBond(p1, p2, [sig_14, eps_14])
+            coul_14_force.addBond(p1, p2, [qq])
+        vdw_force.addExclusion(p1, p2)
+        # electrostatics_force.addExclusion(p1, p2)
+        electrostatics_force.setExceptionParameters(i, p1, p2, 0.0, 0.0, 0.0)
+        # vdw_force.setExceptionParameters(i, p1, p2, 0.0, 0.0, 0.0)
+
+    return vdw_force, electrostatics_force, vdw_14_force, coul_14_force
+
+
+def _apply_switching_function(vdw_handler, force: openmm.NonbondedForce):
+    if not hasattr(force, "setUseSwitchingFunction"):
+        raise ValueError(
+            "Attempting to set switching funcntion on an OpenMM force that does nont support it."
+            f"Passed force of type {type(force)}."
+        )
+    if getattr(vdw_handler, "switch_width", None) is None:
+        force.setUseSwitchingFunction(False)
+    elif vdw_handler.switch_width.m == 0.0:
+        force.setUseSwitchingFunction(False)
+    else:
+        switching_distance = to_openmm_quantity(
+            vdw_handler.cutoff - vdw_handler.switch_width
+        )
+
+        if switching_distance._value < 0:
+            raise UnsupportedCutoffMethodError(
+                "Found a `switch_width` greater than the cutoff distance. It's not clear "
+                "what this means and it's probably invalid. Found "
+                f"switch_width{vdw_handler.switch_width} and cutoff {vdw_handler.cutoff}"
+            )
+
+        force.setUseSwitchingFunction(True)
+        force.setSwitchingDistance(switching_distance)
